# src/docs/swagger/collections.yaml
openapi: 3.0.0
info:
  title: Shopease
  description: The API collection for Shopease
  version: 1.0.0
servers:
  - url: http://localhost:9000/api/v1
    description: Shopease Development Server
  - url: http://api.shopease.shubhampurwar.in/api/v1
    description: Shopease Production Server
...
paths:
...
  /coupons/validity:
    get:
      tags:
        - Coupons
      description: Allows users to check the validity of a coupon
      operationId: checkCouponValidity
      parameters:
        - in: query
          name: couponCode
          required: true
          description: Coupon code that needs to be verified
          schema:
            type: string
      responses:
        '200':
          description: Provided coupon is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      isValid:
                        type: boolean
  /admin/coupons:
    get:
      tags:
        - Coupons
      description: Allows admins to fetch a list of coupons
      operationId: adminGetCoupons
      parameters:
        - in: query
          name: daysUntilExpiration
          description: Number of days until expiration of coupons
          schema:
            type: integer
            format: int32
            min: 1
        - in: query
          name: discountType
          description: Field to filter coupons by their discount type
          schema:
            type: array
            items:
              type: string
              enum:
                - flat
                - percentage
          style: form
          explode: false
        - in: query
          name: status
          description: Field to filter coupons by their status
          schema:
            type: array
            items:
              type: string
              enum:
                - active
                - inactive
                - expired
          style: form
          explode: false
        - in: query
          name: sort
          description: Option to sort coupons by expiry date in asc|desc order
          schema:
            type: string
            enum:
              - expiry_asc
              - expiry_desc
        - $ref: '#/components/parameters/Page'
      responses:
        '200':
          description: A JSON object containing a list of coupons
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Coupon'
          headers:
            X-Total-Count:
              description: Total number of coupons
              schema:
                type: integer

    post:
      tags:
        - Coupons
      description: Allows admins to create a new coupon
      operationId: createCoupon
      requestBody:
        description: A JSON object containing coupon info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponBody'
      responses:
        '201':
          description: A JSON object containing a new coupon
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Coupon'
  /admin/coupons/{couponId}:
    get:
      tags:
        - Coupons
      description: Allows admins to fetch a coupon by ID
      operationId: getCouponById
      parameters:
        - in: path
          name: couponId
          required: true
          description: ID of the coupon to be fetched
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing a coupon
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Coupon'

    put:
      tags:
        - Coupons
      description: Allows admins to update a coupon
      operationId: updateCoupon
      parameters:
        - in: path
          name: couponId
          required: true
          description: ID of the coupon to be updated
          schema:
            type: string
      requestBody:
        description: A JSON object containing new coupon info
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponBody'
      responses:
        '200':
          description: A JSON object containing an updated coupon
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Coupon'

    delete:
      tags:
        - Coupons
      description: Allows admins to delete a coupon
      operationId: deleteCoupon
      parameters:
        - in: path
          name: couponId
          required: true
          description: ID of the coupon to be deleted
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing ID of the deleted coupon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ID'
  /admin/coupons/{couponId}/activate:
    put:
      tags:
        - Coupons
      description: Allows admins to activate a coupon
      operationId: activateCoupon
      parameters:
        - in: path
          name: couponId
          required: true
          description: ID of the coupon to be activated
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an active coupon
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Coupon'
  /admin/coupons/{couponId}/deactivate:
    put:
      tags:
        - Coupons
      description: Allows admins to deactivate a coupon
      operationId: deactivateCoupon
      parameters:
        - in: path
          name: couponId
          required: true
          description: ID of the coupon to be deactivated
          schema:
            type: string
      responses:
        '200':
          description: A JSON object containing an inactive coupon
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Coupon'
...
components:
  parameters:
    Page:
      in: query
      name: page
      description: Number of records to skip
      schema:
        type: integer
        format: int32
        minimum: 1
...
  schemas:
...
    ID:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
...
    Coupon:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        discountType:
          type: string
          enum:
            - flat
            - percentage
        flatDiscount:
          type: integer
        percentageDiscount:
          type: integer
        expiryDate:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - active
            - inactive
            - expired
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        activeStatusLastUpdatedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CouponBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum:
            - flat
            - percentage
        flatDiscount:
          type: integer
        percentageDiscount:
          type: integer
        expiryDate:
          type: string
          format: date-time
...